Add createZipFile helper function
This function will be used to package the artificats file in order to deploy to Elastic Beanstalk.

Just like the `artificats: files` section of `buildspec.yml`, this function accepts glob patterns to add files to the output ZIP file, so we just have to copy the patterns we are already using.